requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow-intel
astunparse==1.6.3
    # via tensorflow-intel
certifi==2024.2.2
    # via requests
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via tqdm
compress-json==1.0.10
    # via sanitize-ml-labels
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flatbuffers==24.3.25
    # via tensorflow-intel
fonttools==4.51.0
    # via matplotlib
gast==0.5.4
    # via tensorflow-intel
google-pasta==0.2.0
    # via tensorflow-intel
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.1
    # via
    #   tensorboard
    #   tensorflow-intel
h5py==3.10.0
    # via
    #   keras
    #   tensorflow-intel
idna==3.6
    # via requests
joblib==1.3.2
    # via scikit-learn
keras==3.1.1
    # via
    #   -r requirements.in
    #   tensorflow-intel
kiwisolver==1.4.5
    # via matplotlib
libclang==18.1.1
    # via tensorflow-intel
markdown==3.6
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via werkzeug
matplotlib==3.8.4
    # via
    #   -r requirements.in
    #   plot-keras-history
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via
    #   keras
    #   tensorflow-intel
namex==0.0.7
    # via keras
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   opt-einsum
    #   pandas
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow-intel
opt-einsum==3.3.0
    # via tensorflow-intel
optree==0.11.0
    # via keras
packaging==24.0
    # via
    #   matplotlib
    #   tensorflow-intel
pandas==2.2.1
    # via
    #   -r requirements.in
    #   plot-keras-history
pillow==10.3.0
    # via matplotlib
plot-keras-history==1.1.38
    # via -r requirements.in
protobuf==4.25.3
    # via
    #   tensorboard
    #   tensorflow-intel
psycopg2==2.9.9
    # via -r requirements.in
pygments==2.17.2
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
requests==2.31.0
    # via
    #   -r requirements.in
    #   tensorflow-intel
rich==13.7.1
    # via keras
sanitize-ml-labels==1.0.51
    # via plot-keras-history
scikit-learn==1.4.1.post1
    # via -r requirements.in
scipy==1.13.0
    # via
    #   -r requirements.in
    #   plot-keras-history
    #   scikit-learn
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow-intel
sqlalchemy==2.0.29
    # via -r requirements.in
support-developer==1.0.5
    # via plot-keras-history
tensorboard==2.16.2
    # via tensorflow-intel
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.16.1
    # via -r requirements.in
tensorflow-intel==2.16.1
    # via tensorflow
tensorflow-io-gcs-filesystem==0.31.0
    # via tensorflow-intel
termcolor==2.4.0
    # via tensorflow-intel
threadpoolctl==3.4.0
    # via scikit-learn
tqdm==4.66.2
    # via -r requirements.in
typing-extensions==4.11.0
    # via
    #   optree
    #   sqlalchemy
    #   tensorflow-intel
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
werkzeug==3.0.2
    # via tensorboard
wheel==0.43.0
    # via astunparse
wrapt==1.16.0
    # via tensorflow-intel

# The following packages are considered to be unsafe in a requirements file:
# setuptools
